Address review comments, make AdminTest deprecated
